+++ src/dep/v0_5_0/match_info.rs
@@ -4,7 +4,7 @@ use serde::Deserialize;
 #[derive(Debug, Deserialize, PartialEq, Clone)]
 pub struct MatchInfo {
     #[serde(rename = "@size")]
-    pub size: usize,
+    pub size: i32,
     #[serde(rename = "Collect")]
     pub collect: Vec<Collect>,
 }
